Automatically backed up by Learn

+++ db/migrate/002_create_costume_stores.rb
@@ -4,7 +4,7 @@ class CreateCostumeStores < ActiveRecord::Migration[5.2]
     create_table :costume_stores do |t|
       t.string :name
       t.string :location
-      t.integer :costume_invetory
+      t.float :costume_invetory
       t.integer :num_of_employees
       t.boolean :still_in_business
       t.datetime :opening_time
